Tower Defense
I use the term tower defense lightly though I think it fits for the most
part. The Main goal of this game was to make minions spawn at certain
points. At the moment enemies spawn endlessly, while the players minions
spawn by clicking a button. It's not very balanced and the current build
folder is not up to date. I will be updating this as soon as possible.